refactor(es6): improvements and fix template issue
General improvements to the code supporting ES2015 features and also
fixes an issue where Gulp was breaking due to a `version` variable not
being defined. Turns out that `gulp-template` gets crazy when trying to
parse the `template/cli.js` file because it does a require for npm
manifest's version property.
Fixed that by just setting an empty string as the value for the `version` key.